move patched ct to root
